Check if number of episodes has been reached
At the end of each episode, the server checks if the number of trials done by the user is higher or equal to the number of episodes needed to finish an experiment. If yes, it forwards to the browser the URL to which the user will be automatically redirected to.